move entry handler to own file and test file
